Move ui elements inside ui folder
