+++ index.html
@@ -405,11 +405,11 @@
       }
     </script>
     <script type="module" src="js/state/main.js"></script>
-    <script type="module" src="js/modals/welcome.js"></script>
-    <script type="module" src="js/modals/input.js"></script>
-    <script type="module" src="js/modals/information.js"></script>
+    <script type="module" src="js/ui/modals/welcome.js"></script>
+    <script type="module" src="js/ui/modals/input.js"></script>
+    <script type="module" src="js/ui/modals/information.js"></script>
     <script type="module" src="js/state/view-selector.js"></script>
     <script type="module" src="js/state/current-file.js"></script>
-    <script type="module" src="js/toggle/switch-deploy.js"></script>
+    <script type="module" src="js/ui/toggle/switch-deploy.js"></script>
   </body>
 </html>
