+++ js/ui/toggle/switch-deploy.js
